feat: доработать компонент Profile
