+++ src/components/Notification/Notification.jsx
@@ -1,7 +1,14 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import './Notification.css';
 
 const Notification = ({ notificationIsOpen, setNotificationIsOpen, errorMessage }) => {
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setNotificationIsOpen(false);
+    }, 3000);
+    return () => clearTimeout(timer);
+  }, [notificationIsOpen]);
+
   return (
     <div className={`notification ${notificationIsOpen && 'notification_active'}`}>
       <p className='notification__message'>{errorMessage}</p>
