+++ src/utils/MainApi.js
@@ -4,7 +4,7 @@ function getResponseData(res, setErrorMessage) {
   if (!res.ok) {
     //получаем ответ от сервера с текстом ошибки, чтобы передать его в попап
     res.text().then((text) => {
-      setErrorMessage(JSON.parse(text).message || JSON.parse(text).error);
+      setErrorMessage(JSON.parse(text).message || 'Произошла ошибка');
     });
     return Promise.reject(`Ошибка: ${res.status}`);
   }
@@ -112,16 +112,20 @@ export function removeMovie(movieID) {
     credentials: 'include',
   });
 }
-export function updateProfile(data) {
-  return request('/users/me', {
-    method: 'PATCH',
-    headers: {
-      'Content-Type': 'application/json',
+export function updateProfile(data, setErrorMessage) {
+  return request(
+    '/users/me',
+    {
+      method: 'PATCH',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      credentials: 'include',
+      body: JSON.stringify({
+        name: data.name,
+        email: data.email,
+      }),
     },
-    credentials: 'include',
-    body: JSON.stringify({
-      name: data.name,
-      email: data.email,
-    }),
-  });
+    setErrorMessage
+  );
 }
